Fix VolumeFilterList unintended warnings and other serialization related issues

+++ voreen/modules/bigdataimageprocessing/volumefilterproperties/gaussianfilterproperties.h
@@ -56,7 +56,7 @@ private:
         int extentY_;
         int extentZ_;
         SamplingStrategyType samplingStrategyType_;
-        int outsideVolumeValue_;
+        float outsideVolumeValue_;
     };
     std::map<int, Settings> instanceSettings_;
 
@@ -64,7 +64,7 @@ private:
     IntProperty extentY_;
     IntProperty extentZ_;
     OptionProperty<SamplingStrategyType> samplingStrategyType_;
-    IntProperty outsideVolumeValue_;
+    FloatProperty outsideVolumeValue_;
 };
 
 }
